:bug: fix(qfa): V 的静态量化 scale 改用模型 dtype
mxfp_c8.py 把 v_cache_scale_float_reciprocal 建成
1 / exp2(exponent).to(model_config.dtype)，即 bf16；脚本给的是 fp32，
aclnnAscendQuantV3 直接 EZ1001 拒绝（x:DT_BFLOAT16 vs scale:DT_FLOAT），
INGRAPH-CACHE 因此从没跑到 QFA 那一步。

同时核对了 reshape_and_cache 的三件事：写 K/V、写 K scale、以及由
save_v_scale_flag 保证只在加载时做一次的写 V scale——图内本来就没有第三件，
脚本拆成两次调用是等价的。

+++ scripts/debug/test_qfa_fullgraph_repro_npu.py
@@ -260,8 +260,11 @@ class InGraphCacheWrite:
         ]
         self.values = [torch.randn_like(k) for k in self.keys]
         # v_cache_scale_float_reciprocal, flat because the cache's V scale is
-        # the neutral 127 -- see Cache.
-        self.v_recip = torch.ones(n * d, dtype=torch.float32, device="npu")
+        # the neutral 127 -- see Cache. bf16, not fp32: mxfp_c8.py builds it as
+        # 1 / exp2(exponent).to(model_config.dtype), and npu_quantize refuses a
+        # scale whose dtype differs from x's ("dtype of input x:DT_BFLOAT16 is
+        # not compatible with scale:DT_FLOAT", EZ1001).
+        self.v_recip = torch.ones(n * d, dtype=torch.bfloat16, device="npu")
         self.slot_buf: torch.Tensor | None = None
 
     def allocate(self, step) -> None:
